Guttok-116 : 테스트 계정 권한 추가

+++ src/main/java/com/app/guttokback/user/application/service/UserService.java
@@ -82,7 +82,10 @@ public class UserService {
     @Transactional
     public void userDelete(String email) {
         User user = findByUserEmail(email);
-        userRepository.delete(user);
+        if(user.getAuthorities().contains(Roles.ROLE_USER)) {
+            userRepository.delete(user);
+        }
+
     }
 
     public UserDetailInfo userDetail(String email) {
